requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
psycopg2-binary
flask-migrate
flask-babel 